remove in log exists only

+++ src/utils/clearLogFile.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import { LOGFILE } from './loggerInfo';
 export const clearLogFile = () => {
-
+    if (fs.existsSync(LOGFILE))
         fs.unlinkSync(LOGFILE);
 }
